Change Integer to BigDecimal

+++ db/init/02-tables.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS assets
     user_id    BIGINT      NOT NULL REFERENCES users (id) ON DELETE CASCADE ON UPDATE RESTRICT,
     symbol     TEXT        NOT NULL,
     type       asset_type  NOT NULL,
-    amount     INT         NOT NULL,
+    amount     DECIMAL(28, 8)        NOT NULL,
     parameters JSONB,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT now(),
